Split off preferences/project properties into their own files with ES6 classes
