+++ main.js
@@ -4,36 +4,7 @@ Builder = {
     Version: "1.23",
     MenuItems: { list: [], run: null, stop: null, fork: null },
     Platform: require("os").type(),
-    PreferencesPath: Electron_App.getPath("userData") + "/GMEdit/config/Builder-preferences.json",
-    PreferencesElement: document.createElement("div"),
-    Preferences: {
-        reuseTab: false,
-        saveCompile: false,
-        stopCompile: false,
-        displayLine: true,
-        forkArguments: "",
-        runtimeSettings: {
-            Stable: {
-                location: process.env.ProgramData + "/GameMakerStudio2/Cache/runtimes/",
-                runtimeList: [],
-                selection: ""
-            },
-            Beta: {
-                location: process.env.ProgramData + "/GameMakerStudio2-Beta/Cache/runtimes/",
-                runtimeList: [],
-                selection: ""
-            }
-        }
-    },
     RuntimeSettings: null,
-    SavePreferences: function() {
-        Electron_FS.writeFileSync(this.PreferencesPath, JSON.stringify(this.Preferences, (k, v) => {
-            return k == "runtimeList" ? undefined : v;
-        }, "    "));
-    },
-    LoadPreferences: function() {
-        return Object.assign(this.Preferences, JSON.parse(Electron_FS.readFileSync(this.PreferencesPath)));
-    },
     LoadKeywords: function(path) {
         Electron_FS.readdirSync(path).forEach((e) => {
             let RuntimeStat = Electron_FS.statSync(path + "/" + e);
@@ -94,29 +65,15 @@ Builder = {
             Electron_Dialog.showErrorBox("builder error", `builder v${Builder.Version} is not supported on your platform (${require("os").type()})`);
             return false;
         }
-
+        
         // Load preferences file
-        if (Electron_FS.existsSync(this.PreferencesPath) == true) {
-            try {
-                this.Preferences = this.LoadPreferences();
-                if (this.Preferences.runtimeLocation != null) { // migration
-                    this.Preferences.runtimeSettings.Stable.location = this.Preferences.runtimeLocation;
-                    delete this.Preferences.runtimeLocation;
-                    this.Preferences.runtimeSettings.Stable.selection = this.Preferences.runtimeSelection;
-                    delete this.Preferences.runtimeSelection;
-                    delete this.Preferences.runtimeList;
-                }
-            } catch(e) {
-                console.error("builder - Failed to read or parse preferences file.");
-            }
-        } else {
-            this.SavePreferences();
-        }
-        for (let [key, val] of Object.entries(this.Preferences.runtimeSettings)) {
+        BuilderPreferences.init();
+        for (let [key, val] of Object.entries(BuilderPreferences.current.runtimeSettings)) {
             this.InitalizeRuntimes(val, key == "Stable");
         }
         
-        Builder.LoadKeywords(this.Preferences.runtimeSettings.Stable.location + this.Preferences.runtimeSettings.Stable.selection);
+        let runtimeSettings = BuilderPreferences.current.runtimeSettings.Stable;
+        Builder.LoadKeywords(runtimeSettings.location + runtimeSettings.selection);
         return true;
     }
 };
@@ -168,93 +125,9 @@ Builder = {
             }
             if (Builder.MenuItems.run == null) return; // probably running in GMLive.js
 
-            // Create preferences menu!
-            let Preferences = $gmedit["ui.Preferences"];
-            for (let [key, set] of Object.entries(Builder.Preferences.runtimeSettings)) {
-                let runtimeGroup = Preferences.addGroup(Builder.PreferencesElement, `Runtime Settings (${key})`);
-                let element, label;
-                
-                element = Preferences.addInput(runtimeGroup, "Runtime Location", set.location, (value) => {
-                    set.location = value; Builder.SavePreferences();
-                });
-                let runtimeLocationInput = element.querySelector("input");
-                label = element.querySelector("label");
-                label.appendChild(document.createTextNode(" ("));
-                label.appendChild(Preferences.createFuncAnchor("Reset", function() {
-                    switch (key) {
-                        case "Stable": set.location = process.env.ProgramData + "/GameMakerStudio2/Cache/runtimes/"; break;
-                        case "Beta": set.location = process.env.ProgramData + "/GameMakerStudio2-Beta/Cache/runtimes/"; break;
-                        default: return;
-                    }
-                    runtimeLocationInput.value = set.location;
-                    Builder.SavePreferences();
-                }));
-                label.appendChild(document.createTextNode(")"));
-                
-                element = Preferences.addDropdown(runtimeGroup, "Current Runtime", set.selection, set.runtimeList, (value) => {
-                    set.selection = value;
-                    Builder.SavePreferences();
-                });
-                let runtimeListSelect = element.querySelector("select");
-                label = element.querySelector("label");
-                label.appendChild(document.createTextNode(" ("));
-                label.appendChild(Preferences.createFuncAnchor("Rescan", function() {
-                    runtimeListSelect.innerHTML = "";
-                    for (let rt of Builder.GetRuntimes(set.location)) {
-                        let option = document.createElement("option");
-                        option.innerHTML = option.value = rt;
-                        runtimeListSelect.appendChild(option);
-                    }
-                    runtimeListSelect.value = set.selection;
-                    Builder.SavePreferences();
-                }));
-                label.appendChild(document.createTextNode(")"));
-            }
+            BuilderPreferences.ready();
+            BuilderProjectProperties.ready();
             
-            let settingsGroup = Preferences.addGroup(Builder.PreferencesElement, "Builder Settings");
-            if (Builder.Platform =="win") {
-                Preferences.addButton(settingsGroup, "Clean Virtual Drives", () => {
-                    let command = require("child_process");
-                    let conf = Builder.DrivesFile;
-                    if (conf.sync()) conf.data = [];
-                    let done = [];
-                    for (let c of conf.data) {
-                        try {
-                            command.execSync(`subst /d ${c}:`);
-                            done.push(c);
-                        } catch(e) {};
-                    }
-                    conf.data = [];
-                    conf.flush();
-                    Electron_Dialog.showMessageBox({type: "info", title: "Builder", message: `Finished cleaning virtual drives (${done.join(", ")}).`});
-                });
-            }
-            Preferences.addInput(settingsGroup, "Fork Arguments", Builder.Preferences.forkArguments, (value) => {
-                Builder.Preferences.forkArguments = value;
-                Builder.SavePreferences();
-            });
-            Preferences.addCheckbox(settingsGroup, "Reuse Output Tab", Builder.Preferences.reuseTab, (value) => {
-                Builder.Preferences.reuseTab = value;
-                Builder.SavePreferences();
-            });
-            Preferences.addCheckbox(settingsGroup, "Save Upon Compile", Builder.Preferences.saveCompile, (value) => {
-                Builder.Preferences.saveCompile = value;
-                Builder.SavePreferences();
-            });
-            Preferences.addCheckbox(settingsGroup, "Stop Upon Compile", Builder.Preferences.stopCompile, (value) => {
-                Builder.Preferences.stopCompile = value;
-                Builder.SavePreferences();
-            });
-            Preferences.addCheckbox(settingsGroup, "Display Line After Fatal Error", Builder.Preferences.displayLine, (value) => {
-                Builder.Preferences.displayLine = value;
-                Builder.SavePreferences();
-            });
-            Preferences.addButton(Builder.PreferencesElement, "Back", () => {
-                Preferences.setMenu(Preferences.menuMain);
-                Builder.SavePreferences();
-            });
-            Preferences.addText(Builder.PreferencesElement, `builder v${Builder.Version} by nommiin`);
-
             // Add ace commands!
             let AceCommands = $gmedit["ace.AceCommands"];
             AceCommands.add({ name: "run", bindKey: {win: "F5", mac: "F5"}, exec: Builder.Run }, "Run");
@@ -335,67 +208,20 @@ Builder = {
                 return result;
             }
             
-            GMEdit.on("preferencesBuilt", (e) => {
-                Preferences.addButton(e.target, "builder Settings", () => {
-                    Preferences.setMenu(Builder.PreferencesElement);
-                });
-            });
-            
-            GMEdit.on("projectPropertiesBuilt", (e) => {
-                let project = e.project;
-                let group = Preferences.addGroup(e.target, "builder Settings");
-                const defaultVersion = "<default>";
-                //
-                let versions = [defaultVersion];
-                for (let [key, set] of Object.entries(Builder.Preferences.runtimeSettings)) {
-                    versions = versions.concat(set.runtimeList);
-                }
-                //
-                let json = project.properties.builderSettings;
-                let ensureJSON = (isDefault) => {
-                    let properties = project.properties;
-                    let json = properties;
-                    if (isDefault && json == null) return null;
-                    if (json == null) json = properties.builderSettings = {};
-                    return json;
-                };
-                Preferences.addDropdown(group,
-                    "Version override",
-                    json?.runtimeVersion ?? defaultVersion,
-                    versions,
-                (v) => {
-                    if (v == defaultVersion) v = null;
-                    let json = ensureJSON(v == null);
-                    if (json == null) return;
-                    json.runtimeVersion = v;
-                    $gmedit["ui.project.ProjectProperties"].save(project, project.properties);
-                });
-                //
-                let argsField = Preferences.addInput(group,
-                    "Fork arguments override",
-                    json?.forkArguments ?? "",
-                (v) => {
-                    if (v == "") v = null;
-                    let json = ensureJSON(v == null);
-                    if (json == null) return;
-                    json.forkArguments = v;
-                    $gmedit["ui.project.ProjectProperties"].save(project, project.properties);
-                }).querySelector("input");
-                argsField.placeholder = Builder.Preferences.forkArguments;
-            });
-            
             function projectOpened() {
                 for (let item of Builder.MenuItems.list) item.enabled = false;
                 let project = $gmedit["gml.Project"].current;
                 if (Builder.ProjectVersion(project) == 2) {
                     Builder.MenuItems.run.enabled = true;
                     let runtime;
+                    const pref = BuilderPreferences.current;
                     if (project.version.name == "v23"
-                        && Builder.Preferences.runtimeSettings.Stable.selection < "runtime-2.3"
-                        && Builder.Preferences.runtimeSettings.Beta.selection != ""
+                        && pref.runtimeSettings.Stable.selection < "runtime-2.3"
+                        && pref.runtimeSettings.Beta.selection != ""
                     ) {
-                        runtime = Builder.Preferences.runtimeSettings.Beta;
-                    } else runtime = Builder.Preferences.runtimeSettings.Stable;
+                        // if runtime is set to 2.2.5 but project uses 2.3, prefer a beta runtime
+                        runtime = pref.runtimeSettings.Beta;
+                    } else runtime = pref.runtimeSettings.Stable;
                     Builder.RuntimeSettings = runtime;
                     Builder.LoadKeywords(runtime.location + runtime.selection);
                 }
